+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/azure
 
+## 4.0.70
+
+### Patch Changes
+
+- Updated dependencies [3867a4e]
+  - @ai-sdk/deepseek@3.0.44
+
+## 4.0.69
+
+### Patch Changes
+
+- Updated dependencies [5ec21a6]
+- Updated dependencies [7469a3b]
+- Updated dependencies [0de8886]
+- Updated dependencies [d5e3024]
+- Updated dependencies [0096850]
+- Updated dependencies [813bb36]
+- Updated dependencies [c43e4b7]
+  - @ai-sdk/openai@4.0.66
+  - @ai-sdk/provider@4.0.14
+  - @ai-sdk/deepseek@3.0.43
+  - @ai-sdk/provider-utils@5.0.40
+
 ## 4.0.68
 
 ### Patch Changes